packages/evals/pyproject.toml
[project]
name = "sidekick-evals"
version = "0.1.0"
description = "Sidekick evaluation suite (test cases, LLM Judge, reports)"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.9",
    "pyyaml>=6.0",
    "rich>=13.9",
    "litellm>=1.52",
    "sidekick-agent",
    "sidekick-tools",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sidekick_evals"]
